requirements.txt
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6-sip==13.10.0
python-dotenv==1.0.0
qasync==0.27.1
pulsefire==0.1.0
pytest